+++ src/main/java/name/divinityunbound/datagen/ModModelProvider.java
@@ -101,6 +101,12 @@ public class ModModelProvider extends FabricModelProvider {
         itemModelGenerator.register(ModItems.SPACE_FORGED_SWORD, Models.HANDHELD);
         itemModelGenerator.register(ModItems.SPACE_FORGED_HOE, Models.HANDHELD);
 
+        itemModelGenerator.register(ModItems.SPACE_TIME_PICKAXE, Models.HANDHELD);
+        itemModelGenerator.register(ModItems.SPACE_TIME_AXE, Models.HANDHELD);
+        itemModelGenerator.register(ModItems.SPACE_TIME_SHOVEL, Models.HANDHELD);
+        itemModelGenerator.register(ModItems.SPACE_TIME_SWORD, Models.HANDHELD);
+        itemModelGenerator.register(ModItems.SPACE_TIME_HOE, Models.HANDHELD);
+
         /* Armor */
         itemModelGenerator.registerArmor((ArmorItem) ModItems.CELESTITE_HELMET);
         itemModelGenerator.registerArmor((ArmorItem) ModItems.CELESTITE_CHESTPLATE);